GlobalGrooveAmount is now double

+++ src/als-tools.core/Entities/LiveProject.cs
@@ -63,7 +63,7 @@ public class LiveProject
     /// <summary>
     /// The project global groove amount
     /// </summary>
-    public int GlobalGrooveAmount { get; set; }
+    public double GlobalGrooveAmount { get; set; }
 
     /// <summary>
     /// The tracks this project contains
